requirements.txt
fastapi
uvicorn
psycopg[binary]
python-dotenv
httpx
paramiko
